requirements.txt
# Python dependencies for SPARK2026 Pose Estimation

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Data loading
h5py>=3.8.0

# Configuration
pyyaml>=6.0

# Visualization & logging
tensorboard>=2.12.0
matplotlib>=3.7.0
tqdm>=4.65.0
opencv-python>=4.8.0
scipy>=1.10.0

# Optional: Weights & Biases logging
# wandb>=0.15.0

# Testing
pytest>=7.3.0
